Node/Edge: add {begin|end}Config()
beginConfig() and endConfig() work reference counted.
Only at the last call of endConfig(), i.e. when the reference
counter is 0 again, the signal changed() is emitted.

This way, one can change multiple properties of a Node
or Edge (e.g. text + position) and delay changed(), so that
finally only a single changed() is emitted instead of multiple
signals.