add SliderSpinBox similar to what Krita has
The SliderSpinBox behaves like a QSpinBox, except that it additionally
shows the value in terms of a progress bar instead of the line edit.
The code is originally from Krita's KisSliderSpinBox, however it was
cleaned up quite a lot. The behavior is the same, though.

CCMAIL: justin@ics.com
CCMAIL: cberger@cberger.net