Implement own UndoManager
This undo manager works similar to the QUndoStack, but it
is more powerful. Whenever you perform a transaction,
you call startTransaction(). When the transaction is finished,
you call finishTransaction(). At this point, all changes
between these two calls are grouped into one undo/redo group.
This is also possible with the QUndoStack (through beginMacro
and endMacro).

However, now comes the difference to QUndoStack: If you want
to cancel a transaction started with startTransaction(), you
can also call cancelTransaction(). In this case, all actions
already performed during the transaction are rolled back and
no undo/redo group is added to the undo/redo stack. This has
several advantages:
1. If a transaction is canceled, the redo items are not lost.
2. You can now cancel a transaction with Escape, which is very
   handy to cancel operations you accidentally started.
3. It allows partial undo/redo of undo items, since we have
   more control over the undo stack.

This undo manager implementation is a mix of QUndoStack and
Kate's KateUndoManager.