Add enum class tikz::EntityType and Entity::entityType()
The EntityType is accessible through the new Uid class.
The Uid class keeps the document-wide unique id as well
as the type of an Entity in terms of the EntityType.
As of now, the Uid is not really needed, but it may take
a more central role in (de-)serialization, since it also
encondes the entity type. This is more robust and allows
to catch errors more early.

Further, Entity::entityType() can be used to static_cast
an entity based on the EntityType. Thus, all Entities
can be managed in one container and quickly converted
to what the respective Entity really is.