Add logging facilities that appear in a dock widget
Using log(), debug(), info(), warn(), error() it is possible to
show a log message to the user in a Message dock widget. This is
useful for debugging as well as notifications for the user. E.g.
trying to use an anchor that is not supported by a shape should
result in a warning saying that an invalid anchor was specified.