Entity: Remove Document parameter, the Uid is enough
In fact, the Uid already contains the associated document.
Passing the document parameter meant passing the document twice.
And passing the document twice meant we need asserts to make sure
the documents are equal. Removing the 2nd document parameter
simplifies this.

As a by-product, we loose the ability to pass the parent to the
final QObject. But this is even intended, since Entities are
by design managed by the Document.