Path: Use a Style with a valid Uid
So far, Path hat its own Style with an unknown Uid (-1), meaning
that the Path Style could not be changed via UndoSetProperty. Now,
the Path's Style also has a valid Uid. This allows to remove the
UndoSetPathStyle undo/redo command. As side effect, UndoDeletePath
and UndoCreatePath can also be deleted, since UndoCreate/DeleteEntity
take care of this.

With this change, we loose the capability of creating an EllipsePath.
A problem to be revisited at a later point in time.

This was already done in 5b734480fad69132685fc3211d7cafcf0d014f65 for nodes.